merge

+++ app/src/main/resources/webapp/src/hmodal/hmodal.js
@@ -1,18 +1,21 @@
-import React from 'react';
-import { useState } from 'react';
+import React, { useState, useEffect } from "react";
 import './hmodal.css';
 import axios from 'axios';
+import ReactPlayer from 'react-player';
 
 axios.defaults.withCredentials = true;
 
 function HModal(props) {
-  const { isOpen, onClose, isNo } = props;
+  const { isOpen, onClose, isNo, userNo } = props;
   const [value, setValue] = useState('');
   
   const [image, setImage] = useState('img/heart.png');
   const [prevImage, setPrevImage] = useState('');
 
   const [data, setData] = React.useState([]);
+  const [url, setUrl] = useState('');
+
+  const [likeCnt, setLikeCnt] = useState();
 
   const handleKeyDown = (event) => {
     if (event.keyCode === 13) {
@@ -20,6 +23,43 @@ function HModal(props) {
       insert();
     }
   };
+
+  // console.log(isNo);
+  // console.log(userNo);
+
+
+
+  axios.get("http://localhost/web/like/cnt/" + isNo)
+    .then(response => {
+      const result = response.data;
+      setLikeCnt(result);
+      // console.log(result);
+    })
+    .catch(exception => {
+
+    });
+
+
+  axios.get("http://localhost/web/like/" + isNo)
+  .then(response => {
+    const result = response.data;
+
+    const data = result.data;
+    // console.log(result.data);
+    if (data === 'false') {
+      // 좋아요를 누르지 않은 상태
+      setImage('img/heart.png');
+    } else if (data === 'true') {
+      // 이미 좋아요를 누른 상태
+      setImage('img/colorheart.png');
+    }
+    // console.log(result);
+
+  })
+  .catch(exception => {
+    alert("입력 오류!");
+    console.log(exception);
+  });
 
   const handleChange = (event) => {
     setValue(event.target.value);
@@ -39,11 +79,9 @@ function HModal(props) {
       }
     )
 		.then(response => {
-      // console.log(response);
 			return response;
 		})
 		.then(result => {
-			// console.log(result);
 			if (result.status == '200') {
         window.location.reload();
 			} else if (result.errorCode == '401') {
@@ -60,32 +98,90 @@ function HModal(props) {
 
   React.useEffect(() => {
     axios
-    .get(
-      "http://localhost/web/replys",
-      {},
-      {
-        params: {
-          boardNo: isNo,
-        },
-      }
-    )
-      .then((response) => setData(response.data.data))
+      .get(
+        "http://localhost/web/replys",
+        {},
+        {
+          params: {
+            boardNo: isNo,
+          },
+        }
+      )
+      .then((response) => {
+        setData(response.data.data);
+        const result = response.data;
+        // console.log(result);
+        // checkOwner2(response.data.data.writerNo);
+      })
       .catch((error) => console.error(error));
 
   }, [isNo]);
 
-
   function handleClick() {
     if (image === 'img/heart.png') {
+      
+      axios
+      .post(
+        "http://localhost/web/like",
+        {},
+        {
+          params: {
+            boardNo: isNo,
+            memberNo: userNo,
+          },
+        }
+      )
+      .then(response => {
+        console.log(response);
+        return response;
+      })
+      .then(result => {
+        if (result.status == '200') {
+        } else if (result.errorCode == '401') {
+        } else {
+          alert('입력 실패!');
+        }
+      })
+      .catch(exception => {
+        alert("입력 오류!");
+        console.log(exception);
+      });
+
       setPrevImage('img/heart.png');
       setImage('img/colorheart.png');
     } else {
+
+      axios
+      .post(
+        "http://localhost/web/like/delete",
+        {},
+        {
+          params: {
+            boardNo: isNo,
+            memberNo: userNo,
+          },
+        }
+      )
+      .then(response => {
+        return response;
+      })
+      .then(result => {
+        if (result.status == '200') {
+          console.log(result);
+        } else if (result.errorCode == '401') {
+        } else {
+          alert('입력 실패!');
+        }
+      })
+      .catch(exception => {
+        alert("입력 오류!");
+        console.log(exception);
+      });
+
       setImage(prevImage);
       setPrevImage('');
     }
   }
-
-  // console.log(isNo);
 
   if (!isOpen) {
     return null;
@@ -94,34 +190,33 @@ function HModal(props) {
   axios.get(`http://localhost/web/boards/${isNo}`)
   .then(response => {
     const result = response.data;
+
     if (result.status === 'failure') {
       alert('게시글을 조회할 수 없습니다.');
-      // location.href = "list.html";
       return;
     }
     
     const board = result.data;
-    //console.log(board);
-    // document.querySelector("input[name='no']").value = board.no;
+
     document.querySelector("input[name='title']").value = board.title;
     document.querySelector("#f-created-date").innerHTML = board.createdDate;
     document.querySelector("#f-writer-name").innerHTML = board.writer.nickname;
     document.querySelector("textarea[name='content']").value = board.content;
-    // document.querySelector("#f-view-count").innerHTML = board.viewCount;
-    
+
     let ul = "";
     board.attachedFiles.forEach(file => {
       // console.log(file);
       if (file.no == 0) return;
       let html = `
-        <li id="li-${file.no}">
-          <a href="https://kr.object.ncloudstorage.com/bitcamp-bucket28/board/${file.filepath}">${file.originalFilename}</a>
-          [<a href="#" onclick="deleteFile(${board.no}, ${file.no}); return false;">삭제</a>] 
-        </li>`;
+      <li id="li-${file.no}">
+        <a href="https://kr.object.ncloudstorage.com/bitcamp-bucket22/board/${file.filepath}">${file.originalFilename}</a>
+        [<a href="#" onClick="deleteFile(${board.no}, ${file.no}); return false;">삭제</a>] 
+      </li>`;
+        setUrl(`${file.filepath}`);
+        // console.log(url);
       ul += html;
     });
-    document.querySelector("#f-files").innerHTML = ul;
-    
+    // document.querySelector("#f-files").innerHTML = ul;
     checkOwner(board.writer.no);
   })
   .catch(error => console.error(error));
@@ -163,20 +258,22 @@ function HModal(props) {
       }
 
     function deleteBtn() {
-      axios.delete("http://localhost/web/boards/" + isNo)
-        .then(response => {
-        // console.log(response.data);
-        if (response.data.status === 'success') {
-        window.location.href='./';
-        } else {
-        alert('삭제 실패!');
+        if (window.confirm('정말로 삭제하시겠습니까?')) {
+          axios.delete("http://localhost/web/boards/" + isNo)
+          .then(response => {
+            const result = response.data;
+            if (result.status === 'success') {
+              alert('게시글이 삭제되었습니다.');
+              window.location.reload();
+            } else {
+              alert('게시글 삭제 실패!');
+            }
+          })
+          .catch(error => console.error(error));
         }
-        })
-        .catch(error => {
-        alert('삭제 중 오류 발생!');
-        console.log(error);
-        });
       }
+
+      
 
       function commentDelete(commentNo) {
         axios
@@ -235,7 +332,6 @@ function HModal(props) {
           return String(num).padStart(2, '0');
         }
       }
-  
 
   return (
     <div className="hmodal-background" onClick={e => {
@@ -250,8 +346,14 @@ function HModal(props) {
           <form id='board-form' method='post' enctype="multipart/form-data">
             <div className='hmodal-view'>
 
-              <div className='hmodal-view-file'>
-              </div>
+            <div className='hmodal-view-file'>
+            <ReactPlayer
+              url={url}
+              controls={true}
+              width='100%'
+              height='100%'
+            />
+            </div>
 
               <div className='hmodal-view-text'>
                 <div className='hmodal-view-header'>
@@ -260,22 +362,23 @@ function HModal(props) {
                       <input type='text' name='title' className='title-text'/>
                     </div>
                     <div>
-                      <span>From</span>
+                      <span className='f-created-date'>From</span>
                       <span id="f-created-date"></span>
                     </div>
                   </div>
                   <div className='heartbox'>
                     <img src={image} className="heart" onClick={handleClick}></img>
+                    <div className="likeCntText">{likeCnt}</div>
                   </div>
                 </div>
 
                 <div className='hmodal-view-body'>
                   <div className='hmodal-view-body-head'>
-                    <span id="f-writer-name"></span>
+                    작성자 : <span id="f-writer-name"></span>
                   </div>
                   <div className='hmodal-view-body-body'>
                     <textarea name='content' rows='10' cols='51' className='content-text'></textarea>
-                    <input type="file" name='files' multiple/>
+                    <input type="file" name='files' multiple className="file-input"/>
                     <ul id="f-files"></ul>
                   </div>
                 </div>
@@ -291,12 +394,12 @@ function HModal(props) {
                             return (
                               <li key={item.no}>
                                 <div className='comment-header'>
-                                  <div>작성자: {item.writer.nickname}</div>
-                                  <div>{item.content ? item.content : "내용없음"}</div>
+                                  <div className="comment-header-writer">작성자 : {item.writer.nickname}</div>
+                                  <div className="comment-header-content">{item.content ? item.content : "내용없음"}</div>
                                   
                                 </div>
                                 <div className='comment-footer'>
-                                  <div>
+                                  <div className="comment-footer-createDate">
                                     {formatDate(item.createdDate)}
                                   </div>
                                   <button id='comment-delete' className='comment-delete' type='button' onClick={() => commentDelete(item.no)}>X</button>
